// src/generated/models/Fastpass_employeesModel.ts
/*!
 *
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Fastpass_employeesstatecode = {
  0: 'Active',
  1: 'Inactive',
} as const;
export type Fastpass_employeesstatecode = keyof typeof Fastpass_employeesstatecode;
export const Fastpass_employeesstatuscode = {
  1: 'Active',
  2: 'Inactive',
} as const;
export type Fastpass_employeesstatuscode = keyof typeof Fastpass_employeesstatuscode;

export interface Fastpass_employeesBase {
  fastpass_currentmilestone?: string;
  fastpass_department?: string;
  fastpass_employeecode?: string;
  fastpass_employeeid: string;
  fastpass_fullname?: string;
  fastpass_journeystatus?: string;
  fastpass_lastactivitydate?: string;
  fastpass_managername?: string;
  fastpass_progresspercentage?: number;
  fastpass_role?: string;
  fastpass_startdate?: string;
  fastpass_team?: string;
  fastpass_userprincipalname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Fastpass_employeesstatecode;
  statuscode?: Fastpass_employeesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}
export interface Fastpass_employees extends Fastpass_employeesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}