// src/generated/models/Fastpass_employeetasksModel.ts
/*!
 *
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Fastpass_employeetasksstatecode = {
  0: 'Active',
  1: 'Inactive',
} as const;
export type Fastpass_employeetasksstatecode = keyof typeof Fastpass_employeetasksstatecode;
export const Fastpass_employeetasksstatuscode = {
  1: 'Active',
  2: 'Inactive',
} as const;
export type Fastpass_employeetasksstatuscode = keyof typeof Fastpass_employeetasksstatuscode;

export interface Fastpass_employeetasksBase {
  fastpass_blockerdescription?: string;
  fastpass_blockerflag?: boolean;
  fastpass_category?: string;
  fastpass_completeddate?: string;
  fastpass_description?: string;
  fastpass_duedate?: string;
  'fastpass_Employee@odata.bind'?: string;
  fastpass_employeetaskid: string;
  fastpass_notes?: string;
  fastpass_recommendedresourceid?: string;
  fastpass_required?: boolean;
  fastpass_status?: string;
  fastpass_taskname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Fastpass_employeetasksstatecode;
  statuscode?: Fastpass_employeetasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}
export interface Fastpass_employeetasks extends Fastpass_employeetasksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  fastpass_blockerflagname?: string;
  fastpass_employeename?: string;
  fastpass_requiredname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  fastpass_employee?: object;
  _fastpass_employee_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}