// src/generated/models/Fastpass_milestonesModel.ts
/*!
 *
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Fastpass_milestonesstatecode = {
  0: 'Active',
  1: 'Inactive',
} as const;
export type Fastpass_milestonesstatecode = keyof typeof Fastpass_milestonesstatecode;
export const Fastpass_milestonesstatuscode = {
  1: 'Active',
  2: 'Inactive',
} as const;
export type Fastpass_milestonesstatuscode = keyof typeof Fastpass_milestonesstatuscode;

export interface Fastpass_milestonesBase {
  fastpass_description?: string;
  fastpass_milestonecode?: string;
  fastpass_milestoneid: string;
  fastpass_milestonename?: string;
  fastpass_tasknames?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Fastpass_milestonesstatecode;
  statuscode?: Fastpass_milestonesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}
export interface Fastpass_milestones extends Fastpass_milestonesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}