// src/generated/models/Fastpass_resourcesModel.ts
/*!
 *
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Fastpass_resourcesstatecode = {
  0: 'Active',
  1: 'Inactive',
} as const;
export type Fastpass_resourcesstatecode = keyof typeof Fastpass_resourcesstatecode;
export const Fastpass_resourcesstatuscode = {
  1: 'Active',
  2: 'Inactive',
} as const;
export type Fastpass_resourcesstatuscode = keyof typeof Fastpass_resourcesstatuscode;

export interface Fastpass_resourcesBase {
  fastpass_description?: string;
  fastpass_relatedtasknames?: string;
  fastpass_resourcecode?: string;
  fastpass_resourceid: string;
  fastpass_resourcename?: string;
  fastpass_type?: string;
  fastpass_url?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Fastpass_resourcesstatecode;
  statuscode?: Fastpass_resourcesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}
export interface Fastpass_resources extends Fastpass_resourcesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}